refactor: extract keyboard and submit-review-flow hooks from Overlay
Overlay.tsx drops from 779 to ~330 lines. The window-capture keydown
effect moves into useOverlayKeyboard (wiring/deps unchanged — modal
ref-plumbing is load-bearing and intentionally untouched), and the
Submit Review / Connect GitHub flow moves into useSubmitReviewFlow.
Overlay itself is now layout + a handful of lifecycle effects.
